refactor(workspace): move constants to tools folder
Move tool-related constants and types from workspace/constants/
to workspace/tools/ where they belong:
- WORKSPACE_TOOLS, WORKSPACE_TOOLS_PREFIX, WorkspaceToolName → tools/constants.ts
- WorkspaceToolConfig, WorkspaceToolsConfig → tools/types.ts

Delete the now-empty constants/ folder.
