+++ packages/core/src/workspace/tools/constants.ts
@@ -1,0 +1,51 @@
+/**
+ * Workspace Tool Constants
+ *
+ * BROWSER-SAFE EXPORTS ONLY
+ *
+ * Constants for workspace tools with no Node.js dependencies,
+ * making it safe to import in browser environments.
+ */
+
+export const WORKSPACE_TOOLS_PREFIX = 'mastra_workspace' as const;
+const TOOLS_NAMESPACE = WORKSPACE_TOOLS_PREFIX;
+
+/**
+ * Workspace tool name constants.
+ * Use these to reference workspace tools by name.
+ *
+ * @example
+ * ```typescript
+ * import { WORKSPACE_TOOLS } from '@mastra/core/workspace';
+ *
+ * if (toolName === WORKSPACE_TOOLS.SANDBOX.EXECUTE_COMMAND) {
+ *   // Handle sandbox execution
+ * }
+ * ```
+ */
+export const WORKSPACE_TOOLS = {
+  FILESYSTEM: {
+    READ_FILE: `${TOOLS_NAMESPACE}_read_file` as const,
+    WRITE_FILE: `${TOOLS_NAMESPACE}_write_file` as const,
+    EDIT_FILE: `${TOOLS_NAMESPACE}_edit_file` as const,
+    LIST_FILES: `${TOOLS_NAMESPACE}_list_files` as const,
+    DELETE: `${TOOLS_NAMESPACE}_delete` as const,
+    FILE_STAT: `${TOOLS_NAMESPACE}_file_stat` as const,
+    MKDIR: `${TOOLS_NAMESPACE}_mkdir` as const,
+  },
+  SANDBOX: {
+    EXECUTE_COMMAND: `${TOOLS_NAMESPACE}_execute_command` as const,
+  },
+  SEARCH: {
+    SEARCH: `${TOOLS_NAMESPACE}_search` as const,
+    INDEX: `${TOOLS_NAMESPACE}_index` as const,
+  },
+} as const;
+
+/**
+ * Type representing any workspace tool name.
+ */
+export type WorkspaceToolName =
+  | (typeof WORKSPACE_TOOLS.FILESYSTEM)[keyof typeof WORKSPACE_TOOLS.FILESYSTEM]
+  | (typeof WORKSPACE_TOOLS.SEARCH)[keyof typeof WORKSPACE_TOOLS.SEARCH]
+  | (typeof WORKSPACE_TOOLS.SANDBOX)[keyof typeof WORKSPACE_TOOLS.SANDBOX];
