+++ packages/core/src/workspace/tools/index.ts
@@ -1,2 +1,4 @@
+export * from './constants';
+export * from './types';
 export * from './tools';
 export * from './tree-formatter';
