+++ packages/core/src/workspace/workspace.ts
@@ -32,7 +32,6 @@
 
 import type { MastraVector } from '../vector';
 
-import type { WorkspaceToolsConfig } from './constants';
 import { WorkspaceError, SearchNotAvailableError } from './errors';
 import type { WorkspaceFilesystem } from './filesystem';
 import type { WorkspaceSandbox } from './sandbox';
@@ -40,6 +39,7 @@ import { SearchEngine } from './search';
 import type { BM25Config, Embedder, SearchOptions, SearchResult, IndexDocument } from './search';
 import type { WorkspaceSkills, SkillsResolver } from './skills';
 import { WorkspaceSkillsImpl, LocalSkillSource } from './skills';
+import type { WorkspaceToolsConfig } from './tools';
 import type { WorkspaceStatus } from './types';
 
 // =============================================================================
